fix: removed webworker implementation since fflate already has webworker.
