+++ packages/common/package.json
@@ -32,7 +32,6 @@
     "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-dts": "^5.3.0",
-    "rollup-plugin-web-worker-loader": "^1.6.1",
     "typescript": "^4.7.4"
   },
   "engines": {
