--- packages/common/types/worker.d.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright 2023 Design Barn Inc.
- */
-
-declare module 'web-worker:*' {
-  const WorkerFactory: new () => Worker;
-  export default WorkerFactory;
-}
